+++ client/src/components/Pagination/PaginationAndNextPage.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
-const SortAndProgress = ({ totalNumberReviews }) => (
+const PaginationAndNextPage = ({ totalNumberReviews }) => (
   <div className="block">
     <span>{totalNumberReviews} out of {totalNumberReviews} reviews</span>
   </div>
 );
 
-export default SortAndProgress;
+export default PaginationAndNextPage;
