+++ package.json
@@ -33,6 +33,7 @@
     "path": "^0.12.7",
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
+    "react-css-modules": "^4.7.11",
     "react-dom": "^16.13.1",
     "webpack-dev-server": "^3.11.0"
   },
